requirements.txt
chromadb==0.5.11
datasets==3.0.1
fastapi==0.111.0
langchain==0.3.3
langchain_chroma==0.1.4
langchain_community==0.3.2
langchain_core==0.3.10
langchain_google_genai==2.0.1
langchain_text_splitters==0.3.0
matplotlib==3.9.2
protobuf==4.25.5
opentelemetry-proto==1.27.0
numpy==1.26.4
opentelemetry-api==1.27.0
pandas==2.2.3
pydantic==2.9.2
pysqlite3_binary==0.5.3.post1
tqdm==4.66.5
umap==0.1.1
jupyterlab==4.2.5
